Fix for Issue 93: Investigate duplicate/mismatched counts in Count Dracula output
I think the previous iteration was too hasty matching movements and over-matched.

CountDracula has a fairly primitive understanding of intersections for right now,
since it uses the streetnames in the excel file name to define the intersection
streets.

(See http://sfcta.github.com/CountDracula/_generated/countdracula.CountsWorkbookParser.html#countdracula.CountsWorkbookParser
for more details)

Thus the incoming street label and outgoing street label aren't necessarily
to be trusted, and the directions (rough as they are) may be more reliable.

This mod is to try to use the streetnames (+ directions, without the primary dir
requirement) first, and then fail back to being based on direction-only, if that
doesn't work out.  The fail back is primarily for when streets change name at
an intersection so the incoming/outgoing labels aren't correct...

Also this is super noisy in the debug log.