Network.addAllMovements(): add functionality:
Additionally, if the incoming link is only one general purpose lane (so the movement is only one lane) and the
outgoing link has multiple options and the movement is turning into a non-gp lane, the movement will shift on
the outgoing link side in order to turn into a general purpose lane.  This is incomplete; it should be true
if the movement is one lane (independent of the lanes in the incoming link) but for the moment that is hard.